requirements.txt
Telethon
requests
urllib3==1.26.6
python-dotenv
beautifulsoup4
html5lib

langchain